Changed typedef for clientProfileProps to profileComponentProps; Combined ClientProfile with UserProfile into component called ProfileComponent; Included clientProfile as prop for ProfileComponent; Allow LoadingAnimation to take in children components for animation; Removed unused labelId for ClientListTable and ClientListTableRow; Switched link to profile to Client column in ClientList and removed Client Profile column
