+++ src/components/Profile/ProfileImageField.jsx
@@ -25,7 +25,7 @@ import { SignedInUserContext } from '../../contexts';
  * @param {profileImageFieldProps} Props - Props used for NewMessage
  * @returns {React.JSX.Element} React component for NewMessage
  */
-const ProfileImageField = ({ loadProfileData }) => {
+const ProfileImageField = ({ loadProfileData, clientProfile }) => {
   const { session } = useSession();
   const { profileData, fetchProfileInfo, removeProfileImage, uploadProfileImage } =
     useContext(SignedInUserContext);
@@ -65,35 +65,36 @@ const ProfileImageField = ({ loadProfileData }) => {
     >
       <Typography color="black">Profile Image: </Typography>
       <Avatar
-        src={profileImg}
+        src={clientProfile ? clientProfile?.profileImg : profileImg}
         alt="PASS profile"
         sx={{ height: '100px', width: '100px', objectFit: 'contain' }}
       />
-      {profileImg ? (
-        <Button
-          variant="outlined"
-          color="error"
-          sx={{ width: '150px' }}
-          onClick={handleRemoveProfileImg}
-          endIcon={<HideImageIcon />}
-        >
-          Remove Img
-        </Button>
-      ) : (
-        <Button
-          variant="outlined"
-          component="label"
-          color="primary"
-          id="input-profile-pic"
-          name="inputProfilePic"
-          onChange={handleProfileImage}
-          endIcon={<ImageIcon />}
-          sx={{ width: '150px' }}
-        >
-          Choose Img
-          <input type="file" hidden accept=".gif, .png, .jpeg, .jpg, .webp" />
-        </Button>
-      )}
+      {!clientProfile &&
+        (profileImg ? (
+          <Button
+            variant="outlined"
+            color="error"
+            sx={{ width: '150px' }}
+            onClick={handleRemoveProfileImg}
+            endIcon={<HideImageIcon />}
+          >
+            Remove Img
+          </Button>
+        ) : (
+          <Button
+            variant="outlined"
+            component="label"
+            color="primary"
+            id="input-profile-pic"
+            name="inputProfilePic"
+            onChange={handleProfileImage}
+            endIcon={<ImageIcon />}
+            sx={{ width: '150px' }}
+          >
+            Choose Img
+            <input type="file" hidden accept=".gif, .png, .jpeg, .jpg, .webp" />
+          </Button>
+        ))}
     </Box>
   );
 };
